+++ blocklets/ai-studio/api/src/store/models/history.ts
@@ -43,6 +43,11 @@ export default class History extends Model<InferAttributes<History>, InferCreati
   declare error?: { message: string };
 
   declare generateStatus?: 'generating' | 'done';
+
+  // 用量上报状态
+  // counted: 已经把这条记录作为**提交点**（但有可能因为上报失败而没有变为 reported）
+  // reported: 已经把上一个**提交点**到这条记录间的 usage 上报至 payment
+  declare usageReportStatus?: null | 'counted' | 'reported';
 }
 
 History.init(
@@ -97,6 +102,9 @@ History.init(
       type: DataTypes.JSON,
     },
     generateStatus: {
+      type: DataTypes.STRING,
+    },
+    usageReportStatus: {
       type: DataTypes.STRING,
     },
   },