+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.1.284 (2024-3-13)
+
+- chore: add pages-kit component dep
+- feat: check user subscription and report usage if needed
+- feat: support config payment for release
+
+## 0.1.283 (2024-3-12)
+
+- fix: use async fs api
+
+## 0.1.282 (2024-3-12)
+
+- feat: support history message datasets api
+
 ## 0.1.281 (2024-3-11)
 
 - fix: remove useless files after project deleted